Change AnimationSequenceBuilder  to struct
