--- Runtime/GamePlay/Playables/PlayableTask.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0fe3fbba4a1814345ae3c9a7e5c2849d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
